Made /me User request part of login process
The User information is now fetched and saved as a cookie just after the
accessTokenJWT is fetched. This will allow the playlist to be created
more easily (it requires the user's info) and will avoid unecessary
extra requests just to display the user's name and profile pic on
multiple pages.