Working Playlist generation!
Added functions for creating and then populating a playlist, and linked
them with the fetch-and-filter library function to complete the core
functionality of Brushma!

I changed the Track interfaces back to using URIs instead of IDs, as I
misread the API docs, and URIs are needed for adding tracks to a
playlist.

There are still some kinks to work out. The first time I tried it on my
entire library, I got some error in the getFilteredLibrary function.
I'll have to try it again and see if that persists or if it was just a
bad moment for the Spotify servers.

Additionally, the playlist description is sometimes applied and
sometimes not, seemingly without reason.