Changing app structure to be multi-page
The app is now divided into several routes
(in addition to the API handlers in '/api'):

'/':
Homepage, including SignInWithSpotify button

'/spotify-redirect':
Page redirected to by Spotify after authorization.
Responsible for fetching the user's access token and profile
information.

'/in':
Page redirected to after signing in.
Shows the signed-in user's info and profile pic.
Contains form for selecting options and a button to begin generating the
playlist.

'/generate':
Page redirected to by form from '/in'.
Fetches and filters library, creates and populates playlist.
Will show results and confirmation of success.